requirements.txt
aws-cdk-lib==2.133.0
awscli==1.29.64
boto3==1.28.64
botocore==1.31.64
constructs==10.3.0
gradio==4.21.0
matplotlib==3.8.0
numpy==1.26.4
pillow==10.2.0
pyyaml==6.0.1
requests==2.31.0